% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes.R
\name{detect.bayes}
\alias{detect.bayes}
\title{Bayesian Stopping Rule for Continuous Data}
\usage{
detect.bayes(GEN, alpha, nulower = NULL, nuupper = NULL,
  score = "hyvarinen", c = 0.5, n = 1000, lenth, thlower = NULL,
  thupper = NULL, GENTH = NULL, ULPTH = NULL, ULP0 = NULL,
  GLP0 = NULL, LLP0 = NULL, ULP1 = NULL, GLP1 = NULL,
  LLP1 = NULL, par0 = NULL, par1 = NULL, lenx = NULL,
  lower0 = NULL, upper0 = NULL, lower1 = NULL, upper1 = NULL)
}
\arguments{
\item{GEN}{A function of time that returns an observation.}

\item{alpha}{A numeric parameter in \code{(0,1)} that controls the probability
of false alarm.}

\item{nulower, nuupper}{Optional nonnegative numerics: The earliest and latest
time of changepoint based on prior belief. The default is \code{nulower=0}
and \code{nuupper=18} which corresponds to the geometric prior distribution
with \code{p=0.1}.}

\item{score}{An optional character specifying the type of score to be used:
The default \code{"hyvarinen"} or the conventional \code{"logarithmic"}.
Can be abbreviated. Case insensitive.}

\item{c, n}{Optional parameters of the Sequentital Monte Carlo algorithm: ESS
threshold \code{0<c<1} and sample size \code{n>0}. Default is \code{c=0.5}
and \code{n=1000}. The resample-move step is triggered whenever the ESS
goes below \code{c*n}.}

\item{lenth}{A positive numeric: The length of the variable of the unknown
parameter in the post-change model.}

\item{thlower, thupper}{Optional numeric vectors of length \code{lenth}: The
lower and upper limits of the unknown parameter in the post-change model.
The defaults are infinite.}

\item{GENTH}{An optional function that takes a sample size and returns a
random sample from the prior distribution of the unknown parameter in the
post-change model. Default is standard normal on the unconstrained
space. Required if \code{ULPTH} is specified.}

\item{ULPTH}{An optional function: The log unnormalized probability function
of the prior distribution of the unknown parameter in the post-change
model. Default is standard normal on the unconstrained space. Required
if \code{GENTH} is specified.}

\item{ULP0, GLP0, LLP0}{Functions of an observation: The log unnormalized
probability function, its gradient and its laplacian for the pre-change
model. If \code{score="hyvarinen"}, either \{\code{GLP0,LLP0}\} or \code{ULP0}
is required. The former is recommended. In the latter case,
\{\code{GLP0,LLP0}\} will be computed via \code{\link[pracma]{grad}} and
\code{\link[pracma]{laplacian}}. If \code{score="logarithmic"}, only
\code{ULP0} is required.}

\item{ULP1, GLP1, LLP1}{Functions of an observation and a numeric parameter:
The log unnormalized probability function, its gradient and its laplacian
for the post-change model. If \code{score="hyvarinen"}, either
\{\code{GLP1,LLP1}\} or \code{ULP1} is required. The former is recommended. In
the latter case, \{\code{GLP1,LLP1}\} will be computed via
\code{\link[pracma]{grad}} and \code{\link[pracma]{laplacian}}. If
\code{score="logarithmic"}, only \code{ULP1} is required.}

\item{par0, par1}{Optional numeric parameters for the pre-change
(\code{par0}) and post-change (\code{par1}) models, except if
\code{score="logarithmic"} that \code{par1} is a function of the unknown
parameter in the post-change model. If \code{score="hyvarinen"}, the positive
tuning parameter with a default of 1. If \code{score="logarithmic"}, the
negative log normalizing constant. If omitted, will be computed via
\code{\link[stats]{integrate}} (if \code{lenx=1}) or
\code{\link[cubature]{hcubature}} (if \code{lenx>1}).}

\item{lenx}{A positive numeric: The length of the variable of an obervation.
Optional if \code{score="hyvarinen"} or if \code{score="logarithmic"} and
\code{par0,par1} are specified.}

\item{lower0, upper0, lower1, upper1}{Optional numeric vectors of length
\code{lenx}: The lower and upper limits of an observation from the
pre-change (\code{lower0,upper0}) and post-change (\code{lower1,upper1})
models. The defaults are infinite.}
}
\value{
A named numeric vector with components
\enumerate{
  \item{\code{t}} {A positive numeric: The stopping time.}
  \item{\code{LER}} {A numeric in \code{(0,1)}: The low ESS rate, i.e., the proportion of iterations that ESS drops below \code{c*n}.}
  \item{\code{AAR}} {A numeric in \code{(0,1)}: The average acceptance rate of the Metropolis-Hastings sampling in the move step. \code{NaN} if ESS never drops below \code{c*n}.}
}
}
\description{
Changepoint detection for continuous data with unknown post-change
distributions using the Bayesian stopping rule.
}
\examples{
##Change from N(0,1) to 2*N(0,1)+1 at t=15.
##Prior knowledge suggests change occurs between 10 and 25.
##The mean and standard deviation of the post-change normal distribution are unknown.

GEN=function(t){ if(15>=t) rnorm(1) else 2*rnorm(1)+1 }
ULP1=function(x,th) -(x-th[1])^2/2/th[2]^2
GLP1=function(x,th) -(x-th[1])/th[2]^2
LLP1=function(x,th) -1/th[2]^2
ULP0=function(x) ULP1(x,c(0,1))
GLP0=function(x) GLP1(x,c(0,1))
LLP0=function(x) LLP1(x,c(0,1))
par0=log(2*pi)/2
par1=function(th) log(2*pi)/2+log(th[2])

#using hyvarinen score
detect.bayes(GEN=GEN,alpha=0.1,nulower=10,nuupper=25,lenth=2,thlower=c(-Inf,0),GLP0=GLP0,LLP0=LLP0,GLP1=GLP1,LLP1=LLP1)

#using log score, normalizing constant unknown
detect.bayes(GEN=GEN,alpha=0.1,nulower=10,nuupper=25,lenth=2,thlower=c(-Inf,0),score="log",ULP0=ULP0,ULP1=ULP1,lenx=1)

#using log score, normalizing constant known
detect.bayes(GEN=GEN,alpha=0.1,nulower=10,nuupper=25,lenth=2,thlower=c(-Inf,0),score="log",ULP0=ULP0,ULP1=ULP1,par0=par0,par1=par1)
}
