% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{detect.bin.stat}
\alias{detect.bin.stat}
\title{Statistic-based Stopping Rule for Binary Data}
\usage{
detect.bin.stat(GEN, alpha, nulower = NULL, nuupper = NULL,
  score = "hyvarinen", ULP0, ULP1, par0 = NULL, par1 = NULL,
  lenx = NULL, tbin = 1)
}
\arguments{
\item{GEN}{A function of time that returns an observation.}

\item{alpha}{A numeric parameter in \code{(0,1)} that controls the probability
of false alarm.}

\item{nulower, nuupper}{Optional nonnegative numerics: The earliest and latest
time of changepoint based on prior belief. The default is \code{nulower=0}
and \code{nuupper=18} which corresponds to the geometric prior distribution
with \code{p=0.1}.}

\item{score}{An optional character specifying the type of score to be used:
The default \code{"hyvarinen"} or the conventional \code{"logarithmic"}.
Can be abbreviated. Case insensitive.}

\item{ULP0, ULP1}{Functions of an observation: The log unnormalized
probability function for the pre-change (\code{ULP0}) and post-change
(\code{ULP1}) models.}

\item{par0, par1}{Optional numeric parameters for the pre-change (\code{par0})
and post-change (\code{par1}) models. If \code{score="hyvarinen"}, the
positive tuning parameter with a default of 1. If
\code{score="logarithmic"}, the negative log normalizing constant. If
omitted, will be computed by summing over the sample space.}

\item{lenx}{A positive numeric: The length of the variable of an
obervation. Optional if \code{score="hyvarinen"} or if
\code{score="logarithmic"} and \code{par0,par1} are specified.}

\item{tbin}{Optional numeric specifying the binary type: The default
\code{tbin=1} representing \{\code{1,0}\} or the alternative \code{tbin=2}
representing \{\code{1,-1}\}.}
}
\value{
A positive numeric: The stopping time.
}
\description{
Changepoint detection for binary data with known post-change distributions
using the statistic-based stopping rule.
}
\examples{
##Change from 3 iid Bernoulli(0.2) to 3 iid Bernoulli(0.8) at t=10.
##Prior knowledge suggests change occurs before 20.

GEN=function(t) { if(10>=t) rbinom(3,1,0.2) else rbinom(3,1,0.8)}
ULP0=function(x) sum(x)*(log(0.2)-log(1-0.2))
ULP1=function(x) sum(x)*(log(0.8)-log(1-0.8))
par0=-3*log(1-0.2)
par1=-3*log(1-0.8)

#using hyvarinen score
detect.bin.stat(GEN=GEN,alpha=0.1,nuupper=20,ULP0=ULP0,ULP1=ULP1)

#using log score. normalizing constant is unknown
detect.bin.stat(GEN=GEN,alpha=0.1,nuupper=20,score="log",ULP0=ULP0,ULP1=ULP1,lenx=3)

#using log score. normalizing constant is known
detect.bin.stat(GEN=GEN,alpha=0.1,nuupper=20,score="log",ULP0=ULP0,ULP1=ULP1,par0=par0,par1=par1)
}
