% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{detect.stat}
\alias{detect.stat}
\title{Statistic-based Stopping Rule for Continuous Data}
\usage{
detect.stat(GEN, alpha, nulower = NULL, nuupper = NULL,
  score = "hyvarinen", ULP0 = NULL, GLP0 = NULL, LLP0 = NULL,
  ULP1 = NULL, GLP1 = NULL, LLP1 = NULL, par0 = NULL,
  par1 = NULL, lenx = NULL, lower0 = NULL, upper0 = NULL,
  lower1 = NULL, upper1 = NULL)
}
\arguments{
\item{GEN}{A function of time that returns an observation.}

\item{alpha}{A numeric parameter in \code{(0,1)} that controls the probability of
false alarm.}

\item{nulower, nuupper}{Optional nonnegative numerics: The earliest and latest
time of changepoint based on prior belief. The default is \code{nulower=0}
and \code{nuupper=18} which corresponds to the geometric prior distribution
with \code{p=0.1}.}

\item{score}{An optional character specifying the type of score to be used:
The default \code{"hyvarinen"} or the conventional \code{"logarithmic"}.
Can be abbreviated. Case insensitive.}

\item{ULP0, GLP0, LLP0, ULP1, GLP1, LLP1}{Functions of an observation: The log
unnormalized probability function, its gradient and its laplacian for the
pre-change (\code{ULP0,GLP0,LLP0}) and post-change (\code{ULP1,GLP1,LLP1}) models. If
\code{score="hyvarinen"}, either \{\code{GLP0,LLP0,GLP1,LLP1}\} or
\{\code{ULP0,ULP1}\} is required. The former is recommended. In the latter case,
\{\code{GLP0,LLP0,GLP1,LLP1}\} will be computed via \code{\link[pracma]{grad}} and
\code{\link[pracma]{laplacian}}. If \code{score="logarithmic"}, only
\{\code{ULP0,ULP1}\} is required.}

\item{par0, par1}{Optional numeric parameters for the pre-change (\code{par0}) and
post-change (\code{par1}) models. If \code{score="hyvarinen"}, the positive
tuning parameter with a default of 1. If \code{score="logarithmic"}, the
negative log normalizing constant. If omitted, will be computed via
\code{\link[stats]{integrate}} (if \code{lenx=1}) or \code{\link[cubature]{hcubature}} (if
\code{lenx>1}).}

\item{lenx}{A positive numeric: The length of the variable of an
obervation. Optional if \code{score="hyvarinen"} or if
\code{score="logarithmic"} and \code{par0,par1} are specified.}

\item{lower0, upper0, lower1, upper1}{Optional numeric vectors of length \code{lenx}:
The lower and upper limits of an observation from the pre-change (\code{lower0,upper0})
and post-change (\code{lower1,upper1}) models. The defaults are infinite.}
}
\value{
A positive numeric: The stopping time.
}
\description{
Changepoint detection for continuous data with known post-change
distributions using the statistic-based stopping rule.
}
\examples{
##Change from N(0,1) to N(1,1) at t=15.
##Prior knowledge suggests change occurs between 10 and 25.

GEN=function(t) { if(15>=t) rnorm(1) else rnorm(1)+1 }
ULP0=function(x) -x^2/2
ULP1=function(x) -(x-1)^2/2
GLP0=function(x) -x
GLP1=function(x) -(x-1)
LLP0=function(x) -1
LLP1=function(x) -1
par0=log(2*pi)/2;par1=par0

#using hyvarinen score
detect.stat(GEN=GEN,alpha=0.1,nulower=10,nuupper=25,GLP0=GLP0,LLP0=LLP0,GLP1=GLP1,LLP1=LLP1)

#using log score. normalizing constant is unknown
detect.stat(GEN=GEN,alpha=0.1,nulower=10,nuupper=25,score="log",ULP0=ULP0,ULP1=ULP1,lenx=1)

#using log score. normalizing constant is known
detect.stat(GEN=GEN,alpha=0.1,nulower=10,nuupper=25,score="log",ULP0=ULP0,ULP1=ULP1,par0=par0,par1=par1)
}
